plotting the redshifts
added a flag so that the redshift of the most likely ions is plotted in a sorta numberline so that we can mybe figure out what winds are happening.

might have to cycle through more temperatures or something to fund better lines.
haven't commented any of the new code but did comment on some of the old.