more thorough search
now the program searches for all of the possible lines within ANY temperature.
which also means now we don't depend on pyatomdb for anything other than downloading the fits files, so all of those issues should go away.
i still need to make it not dependent on the environment variable set by pyatomdb so that we can fully get rid of the library, and i need to comment everything in the new method.